fix(clippy): make the workspace clippy-green under -D warnings
Resolve all `cargo clippy --all-targets --all-features -- -D warnings`
violations so the lint gate passes:

- json/serialize: drop a useless `usize::try_from` on an already-`usize` index.
- vm/lib + runtime/lib: drop needless `&` on `&ExecutionContext` args now that
  NativeCtx borrows the context.
- vm/constructor_fast_path: index-loop -> `iter_mut().take(n).enumerate()`.
- vm/array_prototype: collapse nested `if` into a let-chain.
- jit/builder: `match` -> `matches!`.
- jit/emit: `#[allow(clippy::too_many_arguments)]` on two 8-arg lowering fns.
- jit/baseline (test): drop useless `u64::from` on a `u64` constant.

No behavior change. otter-vm 613 + otter-jit 51 lib tests pass; test262
built-ins/Array 0 fail.

+++ crates/otter-jit/src/baseline.rs
@@ -4785,7 +4785,7 @@ mod tests {
                 Exit::Bailed => panic!("cmp bailed"),
             }
         };
-        let null = (TAG_SPECIAL << 48) | u64::from(super::SPECIAL_NULL);
+        let null = (TAG_SPECIAL << 48) | super::SPECIAL_NULL;
         let obj_a = (super::TAG_PTR_OBJECT << 48) | 0x1234;
         let obj_b = (super::TAG_PTR_OBJECT << 48) | 0x5678;
         assert_eq!(run_cmp(Op::Equal, obj_a, null), f);
